+++ BambooTracker/gui/pattern_editor/pattern_editor.hpp
@@ -34,6 +34,11 @@ public:
 	void freeze();
 	void unfreeze();
 
+	QString getHeaderFont() const;
+	int getHeaderFontSize() const;
+	QString getRowsFont() const;
+	int getRowsFontSize() const;
+	void setFonts(QString headerFont, int headerSize, QString rowsFont, int rowsSize);
 
 signals:
 	void currentTrackChanged(int num);